Temporarily fixed toggleTop()
There should be a check somewhere in place that makes sure the
displayList indexes are valid before (or during) adding them to the
information panel. That way we don't get a blank information panel when
there is something incorrect with displayList (typically on page load
when the url bar has an invalid coin name in it somewhere).