app.js cleanup
Did some overall cleaning of the server side script. Attempted to move
the data retrieval system to another script as was done with routes,
but since we are working with promises it doesn’t pass over data well
through module.exports. Maybe we can look at it again in the future.

Also segregated the bad words object to a .json file so it can be
updated without restarting the server. This also cleaned up a nasty
looking line of code! :)