Information display & cleanup
Made the #filter grow to fill in the rest of the #filterbar height.
Added/removed comments to coinconsole.js for clarity. Added
functionality to change the title text of an information box to green
if positive or red if negative. Added the coin supply to the display.
Adjusted the information box to fit this addition of information,
biggest change went to the title which now helps separate each
information box even more, especially on mobile.